AgentLearn/MiniRAG/requirements.txt
openai
python-dotenv
fastapi
uvicorn
sentence-transformers
faiss-cpu